Split dashboard auth and static UI modules
